+++ crates/nodes/src/rc/empty/info.rs
@@ -1,6 +1,6 @@
 use crate::flags::NodeBitFlags;
 use crate::rc::node::{NodeVariantRc, NodeVariantsRc};
-use crate::types::ZONE_DEFAULT;
+use crate::types::ZONE_ALWAYS;
 use mech3ax_api_types::nodes::rc::Empty;
 use mech3ax_common::{assert_that, Result};
 
@@ -49,7 +49,7 @@ pub(crate) fn assert_variants(node: NodeVariantsRc, offset: usize) -> Result<Nod
     assert_that!("empty flags", const_flags == ALWAYS_PRESENT, offset + 36)?;
     // zero040 (40) already asserted
     assert_that!("empty field 044", node.unk044 in [5, 6], offset + 44)?;
-    assert_that!("empty zone id", node.zone_id in [0, 1, ZONE_DEFAULT], offset + 48)?;
+    assert_that!("empty zone id", node.zone_id >= ZONE_ALWAYS, offset + 48)?;
     // node_type (52) already asserted
     assert_that!("empty data ptr", node.data_ptr == 0, offset + 56)?;
     assert_that!("empty mesh index", node.mesh_index == -1, offset + 60)?;
@@ -81,7 +81,7 @@ pub(crate) fn assert_variants(node: NodeVariantsRc, offset: usize) -> Result<Nod
         name: node.name,
         flags: node.flags.into(),
         unk044: node.unk044,
-        zone_id: node.zone_id,
+        zone_id: node.zone_id as u32, // TODO
         unk116: node.unk116,
         unk140: node.unk140,
         unk164: node.unk164,
@@ -94,7 +94,7 @@ pub(crate) fn make_variants(empty: &Empty) -> NodeVariantsRc {
         name: empty.name.clone(),
         flags: NodeBitFlags::from(&empty.flags),
         unk044: empty.unk044,
-        zone_id: empty.zone_id,
+        zone_id: empty.zone_id as i8, // TODO
         data_ptr: 0,
         mesh_index: -1,
         area_partition: None,