+++ crates/nodes/src/rc/node.rs
@@ -16,7 +16,7 @@ pub struct NodeVariantsRc {
     pub name: String,
     pub flags: NodeBitFlags,
     pub unk044: u32,
-    pub zone_id: u32,
+    pub zone_id: i8,
     pub data_ptr: u32,
     pub mesh_index: i32,
     pub area_partition: Option<AreaPartition>,
@@ -33,7 +33,7 @@ pub struct NodeVariantsRc {
 pub struct NodeVariantLodRc {
     pub name: String,
     pub flags: NodeBitFlags,
-    pub zone_id: u32,
+    pub zone_id: i8,
     pub data_ptr: u32,
     pub has_parent: bool,
     pub parent_array_ptr: u32,
@@ -76,7 +76,9 @@ pub struct NodeRcC {
     flags: Flags,                  // 036
     zero040: u32,                  // 040
     unk044: u32,                   // 044
-    zone_id: u32,                  // 048
+    zone_id: i8,                   // 048
+    pad49: u8,                     // 049
+    pad50: u16,                    // 050
     node_type: NType,              // 052
     data_ptr: Ptr,                 // 056
     mesh_index: i32,               // 060
@@ -128,6 +130,8 @@ fn assert_node(node: NodeRcC, offset: usize) -> Result<(NodeType, NodeVariantsRc
     assert_that!("field 040", node.zero040 == 0, offset + 40)?;
     // unk044 (044) is variable
     // zone_id (048) is variable
+    assert_that!("pad 49", node.pad49 == 0, offset + 49)?;
+    assert_that!("pad 50", node.pad50 == 0, offset + 50)?;
     // node_type (052) see above
     // data_ptr (056) is variable
     // mesh_index (060) is variable
@@ -272,6 +276,8 @@ fn write_variant(
         zero040: 0,
         unk044: variant.unk044,
         zone_id: variant.zone_id,
+        pad49: 0,
+        pad50: 0,
         node_type: node_type.maybe(),
         data_ptr: Ptr(variant.data_ptr),
         mesh_index: variant.mesh_index,
@@ -371,7 +377,9 @@ pub fn assert_node_info_zero(node: &NodeRcC, offset: usize) -> Result<()> {
     assert_that!("field 040", node.zero040 == 0, offset + 40)?;
     assert_that!("field 044", node.unk044 == 0, offset + 44)?;
     assert_that!("zone id", node.zone_id == 0, offset + 48)?;
-    // node type
+    assert_that!("pad 49", node.pad49 == 0, offset + 49)?;
+    assert_that!("pad 50", node.pad50 == 0, offset + 50)?;
+    // node type (52)
     assert_that!("data ptr", node.data_ptr == Ptr::NULL, offset + 56)?;
     assert_that!("mesh index", node.mesh_index == -1, offset + 60)?;
     assert_that!("env data", node.environment_data == 0, offset + 64)?;